+++ lib/uniword/docx.rb
@@ -9,6 +9,7 @@ module Uniword
   module Docx
     autoload :IdAllocator, "#{__dir__}/docx/id_allocator"
     autoload :Part, "#{__dir__}/docx/part"
+    autoload :RawPart, "#{__dir__}/docx/raw_part"
     autoload :ChartPart, "#{__dir__}/docx/chart_part"
     autoload :ImagePart, "#{__dir__}/docx/image_part"
     autoload :HeaderFooterPart, "#{__dir__}/docx/header_footer_part"